requirements.txt
torch==1.7.1
torchvision==0.8.2
pandas==1.2.2
tqdm==4.58.0
pillow==8.1.0
